refactor(edda-cli): split cmd_bridge.rs into cmd_bridge/ directory module (GH-777)
Move the production surface along its existing section markers so bridge
fixes no longer compete for one 3000-line file. Call sites stay on
cmd_bridge::* via re-exports; tests stay in-module with use-line adjustments
only.
