+++ scripts/file-length-ceilings.txt
@@ -12,7 +12,7 @@ crates/edda-bridge-claude/src/digest/tests.rs 3260
 crates/edda-bridge-claude/src/dispatch/tests.rs 3098
 crates/edda-bridge-claude/src/peers/tests.rs 3588
 crates/edda-bridge-claude/src/signals.rs 2252
-crates/edda-cli/src/cmd_bridge.rs 3059
+crates/edda-cli/src/cmd_bridge/tests.rs 1208
 crates/edda-cli/src/cmd_claim.rs 1685
 crates/edda-cli/src/cmd_dispatch.rs 1996
 crates/edda-cli/src/cmd_draft.rs 1705
